fix(loginPage): sets loginPage as homePage
